requirements.txt
# Career Copilot - Unified Requirements
# Core dependencies for local development and production

# =============================================================================
# Core Web Framework
# =============================================================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
starlette>=0.27.0
gunicorn>=21.2.0

# =============================================================================
# Data Validation and Settings
# =============================================================================
pydantic>=2.5.0
pydantic-settings>=2.1.0
pydantic[email]>=2.5.0

# =============================================================================
# Frontend Framework
# =============================================================================
streamlit>=1.29.0

# =============================================================================
# Database and ORM
# =============================================================================
sqlalchemy>=2.0.23
alembic>=1.13.0
psycopg2-binary>=2.9.9
aiosqlite>=0.19.0
asyncpg>=0.29.0

# =============================================================================
# Authentication and Security
# =============================================================================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
bcrypt>=4.1.2
cryptography>=41.0.7
firebase-admin>=6.4.0

# =============================================================================
# AI/LLM Dependencies
# =============================================================================
openai>=1.6.0
anthropic>=0.8.0
groq>=0.4.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langgraph>=0.0.20
langsmith>=0.0.70
tenacity>=8.2.3

# =============================================================================
# Vector Database and Embeddings
# =============================================================================
chromadb>=0.4.18
sentence-transformers>=2.2.2

# =============================================================================
# Caching and Message Queuing
# =============================================================================
redis>=5.0.1
celery>=5.3.4

# =============================================================================
# File Processing and Document Handling
# =============================================================================
python-multipart>=0.0.6
python-magic>=0.4.27
pypdf>=3.17.0
python-docx>=1.1.0
chardet>=5.2.0
aiofiles>=23.2.1
pillow>=10.1.0
pymupdf>=1.23.8

# =============================================================================
# Data Processing and Analytics
# =============================================================================
pandas>=2.1.4
numpy>=1.25.2
plotly>=5.17.0
altair>=5.2.0

# =============================================================================
# HTTP and API Clients
# =============================================================================
requests>=2.31.0
httpx>=0.25.2
urllib3>=2.1.0

# =============================================================================
# WebSocket Support
# =============================================================================
websockets>=12.0
websocket-client>=1.7.0

# =============================================================================
# Configuration and Environment
# =============================================================================
python-dotenv>=1.0.0
pyyaml>=6.0.1

# =============================================================================
# Monitoring and Observability
# =============================================================================
structlog>=23.2.0
prometheus-client>=0.19.0
psutil>=5.9.6
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0

# =============================================================================
# Email and Communication Services
# =============================================================================
sendgrid>=6.10.0
slack-sdk>=3.26.0

# =============================================================================
# Cloud Services
# =============================================================================
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.13.1
google-cloud-logging>=3.8.0
google-cloud-monitoring>=2.16.0

# =============================================================================
# Template Engine and Utilities
# =============================================================================
jinja2>=3.1.2
pyotp>=2.9.0
qrcode>=7.4.2
watchdog>=3.0.0

# =============================================================================
# Development and Testing
# =============================================================================
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0
pytest-cov>=4.1.0
playwright>=1.40.0

# =============================================================================
# Code Quality
# =============================================================================
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0

# =============================================================================
# Additional Utilities
# =============================================================================
python-dateutil>=2.8.2
email-validator>=2.1.0
diskcache>=5.6.3
